requirements.txt
Flask
flask-cors
python-dotenv
boto3